KVM: x86/mmu: Zap shared-only memslots when private attribute changes
Zap all relevant memslots, including shared-only memslots, if the private
memory attribute is being changed.  If userspace converts a range to
private, KVM must zap shared SPTEs to prevent the guest from accessing
the memory as shared.  If userspace converts a range to shared, zapping
SPTEs for shared-only memslots isn't strictly necessary, but doing so
ensures that KVM will install a hugepage mapping if possible, e.g. if a
2MiB range that was mixed is converted to be 100% shared.

Fixes: dcde045383f3 ("KVM: x86/mmu: Handle page fault for private memory")
Link: https://lore.kernel.org/r/20230921203331.3746712-9-seanjc@google.com
Signed-off-by: Sean Christopherson <seanjc@google.com>