KVM: nVMX: Track nested_vmx.posted_intr_nv as a signed int
Track nested_vmx.posted_intr_nv as a signed 32-bit integer instead of an
unsigned 16-bit integer so that it can be passed to kvm_cpu_get_interrupt()
without relying on sign-extension to do the right thing when the vector is
invalid, i.e. when it's -1.

No true functional change intended.

Cc: stable@vger.kernel.org
Reviewed-by: Chao Gao <chao.gao@intel.com>
Link: https://lore.kernel.org/r/20240720000138.3027780-5-seanjc@google.com
Signed-off-by: Sean Christopherson <seanjc@google.com>